Updated KNOTE.py
Edited (edit_menu) + some small functions iNotePAD changed to KNOTE
